requirements-macos.txt
# For Apple Silicon (M1/M2) — pip-style requirements
# This file is a general-purpose list. For the latest PyTorch wheel for your system
# use the official selector: https://pytorch.org/get-started/locally/

numpy
scipy
matplotlib
pandas
scikit-learn
seaborn
statsmodels
jupyterlab

# Recommended deep-learning framework for this course: 
# 1. PyTorch 
# If you explicitly want Apple-accelerated PyTorch builds, follow https://pytorch.org/get-started/locally/
# Example PyTorch packages (may need platform-specific flags):
torch
torchvision
torchaudio
# Pytorch with Metal support (Apple Silicon):
# pip3 install --pre torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/nightly/cpu
# 2. TensorFlow (alternative to PyTorch)
tensorflow-macos
tensorflow-metal # plugin for Apple Silicon
